account panel and post edit/delete

+++ InTouchClient/src/app/shared/components/comment/comment-card/comment-card.component.ts
@@ -126,10 +126,12 @@ export class CommentCardComponent {
 
   onDelete() {
     if(this.comment) {
-      this.commentService.deletePostComment(this.comment?.id).subscribe(success => {
-        if(this.comment)
-        this.deleteComment.emit(this.comment.id)
-      })
+      if(confirm("Do you want to delete the comment?")) {
+        this.commentService.deletePostComment(this.comment?.id).subscribe(success => {
+          if (this.comment)
+            this.deleteComment.emit(this.comment.id)
+        })
+      }
     }
   }
 }
